backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.9.1
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
Flask==2.3.3
Flask-Cors==4.0.0
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.2
openpyxl==3.1.5
pandas==2.3.1
pillow==11.3.0
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-multipart==0.0.5
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
sqlparse==0.5.3
starlette==0.14.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
